moved templating inside pkg dir
