+++ pkg/templating/functions_test.go
